revert commit f30571e7c854910ea413fc6cc6b0589f7bac701b

+++ src/main/java/com/coolerpromc/productiveslimes/compat/jade/JadePlugin.java
@@ -21,7 +21,7 @@ public class JadePlugin implements IWailaPlugin {
         public void appendTooltip(ITooltip iTooltip, EntityAccessor entityAccessor, IPluginConfig iPluginConfig) {
             if (entityAccessor.getEntity() instanceof BaseSlime slime) {
                 int nextDrop = slime.getNextDropTime();
-                iTooltip.add(Component.translatable("tooltip.productiveslimes.next_drop" , (int) Math.ceil((double) nextDrop / 20) + "s"));
+                iTooltip.add(Component.translatable("tooltip.productiveslimes.next_drop" , (int) Math.ceil(nextDrop / 20) + "s"));
             }
         }
